refactor: 시리즈 관리 페이지와 추가 페이지 분리
